floatchat-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
psycopg2-binary
chromadb
